requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.1.1
Django==3.2.8
django-cors-headers==3.13.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
pipreqs==0.4.11
pytz==2021.3
requests==2.28.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzdata==2022.6
urllib3==1.26.12
xlwt==1.3.0
yarg==0.1.9
